Decomposed email sender service
