+++ cmd/gses2-app/main.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	"gses2-app/internal/controllers"
+	"gses2-app/internal/email"
 	"gses2-app/internal/services"
 	"gses2-app/pkg/config"
 )
@@ -20,7 +21,7 @@ func main() {
 	httpClient := &http.Client{Timeout: time.Second * 10}
 	exchangeRateService := services.NewExchangeRateService(httpClient)
 	emailSubscriptionService := services.NewEmailSubscriptionService("./storage.csv")
-	emailSenderService := services.NewEmailSenderService()
+	emailSenderService := email.NewSenderService()
 
 	controller := controllers.NewAppController(
 		exchangeRateService,
